requirements.txt
python-binance
numpy
ta
websocket_client